move most objectseeker methods to the abstract base class
