+++ art/estimators/certification/object_seeker/pytorch.py
@@ -32,13 +32,12 @@ else:
     from typing_extensions import Literal
 
 import numpy as np
-from sklearn.cluster import DBSCAN
 from tqdm import tqdm
 
-from art.config import ART_NUMPY_DTYPE
+from art.estimators.certification.object_seeker.object_seeker import ObjectSeekerMixin
 from art.estimators.object_detection import ObjectDetectorMixin, PyTorchObjectDetector, PyTorchFasterRCNN, PyTorchYolo
 from art.estimators.pytorch import PyTorchEstimator
-from art.utils import intersection_over_union, intersection_over_area, non_maximum_suppression
+from art.utils import intersection_over_area, non_maximum_suppression
 
 if TYPE_CHECKING:
     # pylint: disable=C0412
@@ -51,7 +50,7 @@ if TYPE_CHECKING:
 logger = logging.getLogger(__name__)
 
 
-class PyTorchObjectSeeker(ObjectDetectorMixin, PyTorchEstimator):
+class PyTorchObjectSeeker(ObjectSeekerMixin, ObjectDetectorMixin, PyTorchEstimator):
     """
     Implementation of the ObjectSeeker certifiable robust defense applied to object detection models.
     The original implementation is https://github.com/inspire-group/ObjectSeeker
@@ -96,6 +95,40 @@ class PyTorchObjectSeeker(ObjectDetectorMixin, PyTorchEstimator):
         device_type: str = "gpu",
         verbose: bool = False,
     ):
+        """
+        Create an ObjectSeeker classifier.
+
+        :param model: Object detection model. The output of the model is `List[Dict[str, torch.Tensor]]`,
+                      one for each input image. The fields of the Dict are as follows:
+
+                      - boxes [N, 4]: the boxes in [x1, y1, x2, y2] format, with 0 <= x1 < x2 <= W and
+                        0 <= y1 < y2 <= H.
+                      - labels [N]: the labels for each image.
+                      - scores [N]: the scores of each prediction.
+        :param input_shape: The shape of one input sample.
+        :param optimizer: The optimizer for training the classifier.
+        :param clip_values: Tuple of the form `(min, max)` of floats or `np.ndarray` representing the minimum and
+               maximum values allowed for features. If floats are provided, these will be used as the range of all
+               features. If arrays are provided, each value will be considered the bound for a feature, thus
+               the shape of clip values needs to match the total number of features.
+        :param channels_first: Set channels first or last.
+        :param preprocessing_defences: Preprocessing defence(s) to be applied by the classifier.
+        :param postprocessing_defences: Postprocessing defence(s) to be applied by the classifier.
+        :param preprocessing: Tuple of the form `(subtrahend, divisor)` of floats or `np.ndarray` of values to be
+               used for data preprocessing. The first value will be subtracted from the input. The input will then
+               be divided by the second one.
+        :param attack_losses: Tuple of any combination of strings of loss components: 'loss_classifier', 'loss_box_reg',
+                              'loss_objectness', and 'loss_rpn_box_reg'.
+        :param detector_type: The type of object detector being used: 'YOLO' | 'Faster-RCNN'
+        :param num_lines: The number of divisions both vertically and horizontally to make masked predictions.
+        :param confidence_threshold: The confidence threshold to discard bounding boxes.
+        :param iou_threshold: The IoU threshold to discard overlapping bounding boxes.
+        :param prune_threshold: The IoA threshold for pruning and duplicated bounding boxes.
+        :param epsilon: The maximum distance between bounding boxes when merging using DBSCAN.
+        :param device_type: Type of device to be used for model and tensors, if `cpu` run on CPU, if `gpu` run on GPU
+                            if available otherwise run on CPU.
+        :param verbose: Show progress bars.
+        """
         super().__init__(
             model=model,
             clip_values=clip_values,
@@ -104,18 +137,17 @@ class PyTorchObjectSeeker(ObjectDetectorMixin, PyTorchEstimator):
             postprocessing_defences=postprocessing_defences,
             preprocessing=preprocessing,
             device_type=device_type,
+            num_lines=num_lines,
+            confidence_threshold=confidence_threshold,
+            iou_threshold=iou_threshold,
+            prune_threshold=prune_threshold,
+            epsilon=epsilon,
         )
 
         self._input_shape = input_shape
         self._optimizer = optimizer
         self._attack_losses = attack_losses
-
         self.detector_type = detector_type
-        self.num_lines = num_lines
-        self.confidence_threshold = confidence_threshold
-        self.iou_threshold = iou_threshold
-        self.prune_threshold = prune_threshold
-        self.epsilon = epsilon
         self.verbose = verbose
 
         if detector_type == "YOLO":
@@ -169,6 +201,11 @@ class PyTorchObjectSeeker(ObjectDetectorMixin, PyTorchEstimator):
 
     @property
     def model(self) -> "torch.nn.Module":
+        """
+        Return the model.
+
+        :return: The model.
+        """
         return self._model
 
     @property
@@ -206,8 +243,8 @@ class PyTorchObjectSeeker(ObjectDetectorMixin, PyTorchEstimator):
         :return: Current used device.
         """
         return self._device
-
-    def predict(self, x: np.ndarray, batch_size: int = 128, **kwargs) -> List[Dict[str, np.ndarray]]:
+    
+    def _predict_classifier(self, x: np.ndarray, batch_size: int = 128, **kwargs) -> List[Dict[str, np.ndarray]]:
         """
         Perform prediction for a batch of inputs.
 
@@ -220,170 +257,7 @@ class PyTorchObjectSeeker(ObjectDetectorMixin, PyTorchEstimator):
                  - labels [N]: the labels for each image
                  - scores [N]: the scores or each prediction.
         """
-        predictions = []
-
-        for x_i in tqdm(x, desc="ObjectSeeker", disable=not self.verbose):
-            base_preds = self.detector.predict(x_i[np.newaxis], batch_size=batch_size, **kwargs)[0]
-            filtered_base_preds = non_maximum_suppression(
-                base_preds, iou_threshold=self.iou_threshold, confidence_threshold=self.confidence_threshold
-            )
-
-            masked_preds = self._masked_predictions(x_i, batch_size=batch_size, **kwargs)
-            filtered_masked_preds = non_maximum_suppression(
-                masked_preds, iou_threshold=self.iou_threshold, confidence_threshold=self.confidence_threshold
-            )
-
-            pruned_preds = self._prune_boxes(filtered_masked_preds, filtered_base_preds)
-            unionized_preds = self._unionize_clusters(pruned_preds)
-
-            preds = {
-                "boxes": np.concatenate([filtered_base_preds["boxes"], unionized_preds["boxes"]]),
-                "labels": np.concatenate([filtered_base_preds["labels"], unionized_preds["labels"]]),
-                "scores": np.concatenate([filtered_base_preds["scores"], unionized_preds["scores"]]),
-            }
-
-            predictions.append(preds)
-
-        return predictions
-
-    def _masked_predictions(self, x_i: np.ndarray, batch_size: int = 128, **kwargs) -> Dict[str, np.ndarray]:
-        x_mask = np.repeat(x_i[np.newaxis, ...], self.num_lines * 4, axis=0)
-
-        if self.channels_first:
-            height = self.input_shape[1]
-            width = self.input_shape[2]
-        else:
-            height = self.input_shape[0]
-            width = self.input_shape[1]
-            x_mask = np.transpose(x_mask, (0, 3, 1, 2))
-
-        idx = 0
-
-        # Left masks
-        for k in range(1, self.num_lines + 1):
-            boundary = int(width / (self.num_lines + 1) * k)
-            x_mask[idx, :, :, :boundary] = 0
-            idx += 1
-
-        # Right masks
-        for k in range(1, self.num_lines + 1):
-            boundary = width - int(width / (self.num_lines + 1) * k)
-            x_mask[idx, :, :, boundary:] = 0
-            idx += 1
-
-        # Top masks
-        for k in range(1, self.num_lines + 1):
-            boundary = int(height / (self.num_lines + 1) * k)
-            x_mask[idx, :, :boundary, :] = 0
-            idx += 1
-
-        # Bottom masks
-        for k in range(1, self.num_lines + 1):
-            boundary = height - int(height / (self.num_lines + 1) * k)
-            x_mask[idx, :, boundary:, :] = 0
-            idx += 1
-
-        if not self.channels_first:
-            x_mask = np.transpose(x_mask, (0, 2, 3, 1))
-
-        masked_preds = self.detector.predict(x=x_mask, batch_size=batch_size, **kwargs)
-        filtered_preds = [
-            non_maximum_suppression(
-                pred, iou_threshold=self.iou_threshold, confidence_threshold=self.confidence_threshold
-            )
-            for pred in masked_preds
-        ]
-
-        # Merge predictions
-        boxes = np.concatenate([pred["boxes"] for pred in filtered_preds])
-        labels = np.concatenate([pred["labels"] for pred in filtered_preds])
-        scores = np.concatenate([pred["scores"] for pred in filtered_preds])
-
-        merged_predictions = {
-            "boxes": boxes,
-            "labels": labels,
-            "scores": scores,
-        }
-
-        return merged_predictions
-
-    def _prune_boxes(
-        self, masked_preds: Dict[str, np.ndarray], base_preds: Dict[str, np.ndarray]
-    ) -> Dict[str, np.ndarray]:
-        masked_boxes = masked_preds["boxes"]
-        masked_labels = masked_preds["labels"]
-        masked_scores = masked_preds["scores"]
-
-        base_boxes = base_preds["boxes"]
-        base_labels = base_preds["labels"]
-
-        keep_indices = []
-        for idx, (masked_box, masked_label) in enumerate(zip(masked_boxes, masked_labels)):
-            keep = True
-            for (base_box, base_label) in zip(base_boxes, base_labels):
-                if masked_label == base_label:
-                    ioa = intersection_over_area(masked_box, base_box)
-                    if ioa >= self.prune_threshold:
-                        keep = False
-                        break
-
-            if keep:
-                keep_indices.append(idx)
-
-        pruned_preds = {
-            "boxes": masked_boxes[keep_indices],
-            "labels": masked_labels[keep_indices],
-            "scores": masked_scores[keep_indices],
-        }
-        return pruned_preds
-
-    def _unionize_clusters(self, masked_preds: Dict[str, np.ndarray]) -> Dict[str, np.ndarray]:
-        boxes = masked_preds["boxes"]
-        labels = masked_preds["labels"]
-        scores = masked_preds["scores"]
-
-        unionized_boxes = []
-        unionized_labels = []
-        unionized_scores = []
-
-        unique_labels = np.unique(labels)
-        for label in unique_labels:
-            mask = labels == label
-            selected_boxes = boxes[mask]
-            selected_scores = scores[mask]
-
-            areas = (selected_boxes[:, 2] - selected_boxes[:, 0]) * (selected_boxes[:, 3] - selected_boxes[:, 1])
-            rb = np.minimum(selected_boxes[:, None, 2:], selected_boxes[:, 2:])
-            lt = np.maximum(selected_boxes[:, None, :2], selected_boxes[:, :2])
-            pairwise_intersection = np.prod(np.clip(rb - lt, 0, None), axis=2)
-            pairwise_ioa = pairwise_intersection / (areas[:, None])
-            distances = 1 - np.maximum(pairwise_ioa, pairwise_ioa.T)
-
-            dbscan = DBSCAN(eps=0.1, min_samples=1, metric="precomputed")
-            clusters = dbscan.fit_predict(distances)
-            num_clusters = np.max(clusters) + 1
-
-            for cluster in range(num_clusters):
-                clustered_boxes = selected_boxes[clusters == cluster]
-
-                clustered_box = [
-                    np.min(clustered_boxes[:, 0]),
-                    np.min(clustered_boxes[:, 1]),
-                    np.max(clustered_boxes[:, 2]),
-                    np.max(clustered_boxes[:, 3]),
-                ]
-                clustered_score = np.max(selected_scores)
-
-                unionized_boxes.append(clustered_box)
-                unionized_labels.append(label)
-                unionized_scores.append(clustered_score)
-
-        unionized_predictions = {
-            "boxes": np.array(unionized_boxes),
-            "labels": np.array(unionized_labels),
-            "scores": np.array(unionized_scores),
-        }
-        return unionized_predictions
+        return self.detector.predict(x=x, batch_size=batch_size, **kwargs)
 
     def fit(  # pylint: disable=W0221
         self,
